@@ -2,7 +2,7 @@ ActiveAdmin.register Billing::InvoiceTemplate, as: 'InvoiceTemplate' do
   menu parent: "Billing", label: "Invoice templates", priority: 40
   config.batch_actions = false
   actions :all #:index,:create, :new, :destroy, :delete, :edit, :update
-  before_filter :left_sidebar!
+  before_action :left_sidebar!
 
   permit_params :name, :template_file
 
