@@ -12,7 +12,7 @@ ActiveAdmin.register RealtimeData::ActiveNode do
   filter :name #, as: :select,  collection: proc { SbcNode.pluck(:name)}
   filter :pop
 
-  # before_filter do
+  # before_action do
   #   @index_partial = 'shared/active_calls_chart'
   #   @index_partial_title = 'Active Calls Chart'
   #end
