@@ -16,7 +16,7 @@ ActiveAdmin.register Report::Realtime::TerminationDistribution do
          as: :select, collection: proc { Contractor.select(:id, :name).reorder(:name) },
          input_html: {class: 'chosen'}
 
-  before_filter only: [:index] do
+  before_action only: [:index] do
     params[:q] ||= {}
     if params[:q][:time_interval_eq].blank?
       params[:q][:time_interval_eq] = Report::Realtime::Base::DEFAULT_INTERVAL
