@@ -36,7 +36,7 @@ ActiveAdmin.register CdrExport, as: 'CDR Export' do
 
   controller do
     def build_new_resource
-      build_params = resource_params[0]
+      build_params = resource_params[0].to_h
       return super unless build_params.any?
       #clean up empty character
       build_params['fields'].delete('')
