@@ -1,7 +1,7 @@
 require 'base64'
 
 class Api::RestController < ApiController
-  skip_before_action :verify_authenticity_token
+  skip_before_action :verify_authenticity_token, raise: false
 
   respond_to :json
   rescue_from ActiveRecord::RecordNotFound, with: :render_404
