@@ -36,14 +36,14 @@ describe Api::Rest::Admin::DialpeersController, type: :controller do
     let!(:dialpeer) { create :dialpeer }
 
     context 'when dialpeer exists' do
-      before { get :show, id: dialpeer.to_param }
+      before { get :show, params: { id: dialpeer.to_param } }
 
       it { expect(response.status).to eq(200) }
       it { expect(response_data['id']).to eq(dialpeer.id.to_s) }
     end
 
     context 'when dialpeer does not exist' do
-      before { get :show, id: dialpeer.id + 10 }
+      before { get :show, params: { id: dialpeer.id + 10 } }
 
       it { expect(response.status).to eq(404) }
       it { expect(response_data).to eq(nil) }
@@ -57,9 +57,11 @@ describe Api::Rest::Admin::DialpeersController, type: :controller do
     let(:routing_group) { create :routing_group }
 
     before do
-      post :create, data: { type: 'dialpeers',
-                            attributes: attributes,
-                            relationships: relationships }
+      post :create, params: {
+        data: { type: 'dialpeers',
+                attributes: attributes,
+                relationships: relationships }
+      }
     end
     context 'when attributes are valid' do
       let(:attributes) do
@@ -98,10 +100,12 @@ describe Api::Rest::Admin::DialpeersController, type: :controller do
 
   describe 'PUT update' do
     let!(:dialpeer) { create :dialpeer }
-    before { put :update, id: dialpeer.to_param, data: { type: 'dialpeers',
-                                                         id: dialpeer.to_param,
-                                                         attributes: attributes,
-                                                         relationships: relationships} }
+    before { put :update, params: {
+      id: dialpeer.to_param, data: { type: 'dialpeers',
+                                     id: dialpeer.to_param,
+                                     attributes: attributes,
+                                     relationships: relationships}
+    } }
 
     context 'when attributes are valid' do
       let(:attributes) { { 'next-interval': 90 } }
@@ -123,7 +127,7 @@ describe Api::Rest::Admin::DialpeersController, type: :controller do
   describe 'DELETE destroy' do
     let!(:dialpeer) { create :dialpeer }
 
-    before { delete :destroy, id: dialpeer.to_param }
+    before { delete :destroy, params: { id: dialpeer.to_param } }
 
     it { expect(response.status).to eq(204) }
     it { expect(Dialpeer.count).to eq(0) }
