@@ -25,14 +25,14 @@ describe Api::Rest::Admin::PaymentsController, type: :controller do
     let!(:payment) { create :payment }
 
     context 'when payment exists' do
-      before { get :show, id: payment.to_param }
+      before { get :show, params: { id: payment.to_param } }
 
       it { expect(response.status).to eq(200) }
       it { expect(response_data['id']).to eq(payment.id.to_s) }
     end
 
     context 'when payment does not exist' do
-      before { get :show, id: payment.id + 10 }
+      before { get :show, params: { id: payment.id + 10 } }
 
       it { expect(response.status).to eq(404) }
       it { expect(response_data).to eq(nil) }
@@ -41,9 +41,11 @@ describe Api::Rest::Admin::PaymentsController, type: :controller do
 
   describe 'POST create' do
     before do
-      post :create, data: { type: 'payments',
-                            attributes: attributes,
-                            relationships: relationships }
+      post :create, params: {
+        data: { type: 'payments',
+                attributes: attributes,
+                relationships: relationships }
+      }
     end
 
     context 'when attributes are valid' do
