@@ -16,7 +16,7 @@ describe Api::Rest::Admin::Routing::RoutingTagModesController, type: :controller
   end
 
   describe 'GET show' do
-    before { get :show, id: record.id }
+    before { get :show, params: { id: record.id } }
 
     it 'receive expected fields' do
       expect(response_data.deep_symbolize_keys).to a_hash_including(
@@ -31,8 +31,8 @@ describe Api::Rest::Admin::Routing::RoutingTagModesController, type: :controller
 
   describe 'POST create' do
     subject do
-      post :create, data: { type: resource_type,
-                            attributes: { name: 'Name_1' } }
+      post :create, params: { data: { type: resource_type,
+                            attributes: { name: 'Name_1' } } }
     end
 
     it { expect { subject }.to raise_error(ActionController::UrlGenerationError) }
@@ -40,9 +40,9 @@ describe Api::Rest::Admin::Routing::RoutingTagModesController, type: :controller
 
   describe 'PUT update' do
     subject do
-      put :update, id: record.to_param, data: { type: resource_type,
+      put :update, params: { id: record.to_param, data: { type: resource_type,
                                                 id: record.id.to_i,
-                                                attributes: { name: 'Update name_1' } }
+                                                attributes: { name: 'Update name_1' } } }
     end
 
     it { expect { subject }.to raise_error(ActionController::UrlGenerationError) }
@@ -51,7 +51,7 @@ describe Api::Rest::Admin::Routing::RoutingTagModesController, type: :controller
   describe 'DELETE destroy' do
     before { record }
 
-    subject { delete :destroy, id: record.to_param }
+    subject { delete :destroy, params: { id: record.to_param } }
 
     it { expect { subject }.to raise_error(ActionController::UrlGenerationError) }
   end
