@@ -31,14 +31,14 @@ describe Api::Rest::Admin::RoutingGroupsController, type: :controller do
     let!(:routing_group) { create :routing_group }
 
     context 'when routing_group exists' do
-      before { get :show, id: routing_group.to_param }
+      before { get :show, params: { id: routing_group.to_param } }
 
       it { expect(response.status).to eq(200) }
       it { expect(response_data['id']).to eq(routing_group.id.to_s) }
     end
 
     context 'when routing_group does not exist' do
-      before { get :show, id: routing_group.id + 10 }
+      before { get :show, params: { id: routing_group.id + 10 } }
 
       it { expect(response.status).to eq(404) }
       it { expect(response_data).to eq(nil) }
@@ -46,7 +46,9 @@ describe Api::Rest::Admin::RoutingGroupsController, type: :controller do
   end
 
   describe 'POST create' do
-    before { post :create, data: { type: 'routing-groups', attributes: attributes } }
+    before { post :create, params: {
+      data: { type: 'routing-groups', attributes: attributes }
+    } }
 
     context 'when attributes are valid' do
       let(:attributes) { { name: 'name' } }
@@ -65,9 +67,11 @@ describe Api::Rest::Admin::RoutingGroupsController, type: :controller do
 
   describe 'PUT update' do
     let!(:routing_group) { create :routing_group }
-    before { put :update, id: routing_group.to_param, data: { type: 'routing-groups',
-                                                              id: routing_group.to_param,
-                                                              attributes: attributes } }
+    before { put :update, params: {
+      id: routing_group.to_param, data: { type: 'routing-groups',
+                                          id: routing_group.to_param,
+                                          attributes: attributes }
+    } }
 
     context 'when attributes are valid' do
       let(:attributes) { { name: 'name' } }
@@ -87,7 +91,7 @@ describe Api::Rest::Admin::RoutingGroupsController, type: :controller do
   describe 'DELETE destroy' do
     let!(:routing_group) { create :routing_group }
 
-    before { delete :destroy, id: routing_group.to_param }
+    before { delete :destroy, params: { id: routing_group.to_param } }
 
     it { expect(response.status).to eq(204) }
     it { expect(RoutingGroup.count).to eq(0) }
